requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
fastapi==0.112.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
openai==1.43.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.4
tqdm==4.66.5
typing_extensions==4.12.2
uvicorn==0.30.6
